Some updates to fish shell.
Found out that fish can autocomplete from man pages so run that command
on initial install. May want to run this on upgrade also but right now
it does not have a lot of value as it is not common to add new commands.

Also found fish_add_path function which handles ensuring the PATH
variable does not get continuously lengthened, making the config files
that bit neater.

Have also set the bat commands default colour scheme to be a bit closer
to the colour themes used in other programs.