Remove nerdtree in favour of dirvish
Always had difficulty with the hierarchical notion of nerd tree and
navigating it. Then found this was a pretty common issue people broadly
discussed many years back. Dirvish provides a similar but better
inteface than the built in file explorer in vim which is easier to
navigate with as we treat it as just another buffer rather than an
individual file explorer.