A few vim improvements.
Removed the auto save plugin as this can already be done as part of the
workspace plugin. No need for them both.

Improved airline's tabline extension so can now quickly navigate between
the tabs seen at the top via the numbers shown on them.

Fixed dirvish plugin which added a mapping for <C-p> which overrode my
current workflow. Now when this opens on a new directory I am able to
access files.