Remove some vim plugins.
A few plugins are not used. The fish plugin is already pulled in using
vim-polyglot. Alos vim-polyglot pulls in its own version of vim-sensible
so removing that one also.

When reading up on the fish plugin it actually recommends to use a POSIX
compliant shell so have set the shell to bash instead of fish. Also
means that all my fish config start up will not affect any commands or
plugins.

Also do not use the terminal emulator so I have removed that part of my
configuration.