Lots of bootstrap fixes
Just ran the bootstrap sequence on a new linux install and had some
teething issues that have been addressed. The main issue was around
ensuring that fish was correctly installed and set as default shell. To
do this needed to use apt to install rather than linux brew.

Linux brew also had many issues not being on the path so have fixed that
so that it does not try install constantly.