Add SDK man and fix snap installs.
SDK man is what is used for managing SDK and other tooling around the
JVM. This allows management of different versions if required and is
more frequently updated than apt repositories.

The snap installs now perform the check prior to calling the install
making the bootstrap script faster as it does not continually request
for a password.