Migrate to native LSP.
Neovim 0.5 version just became stable so I have moved across to using
the native version over CoC assuming that it will eventually be better
to work with being native, particularly with plugin ecosystem.

This means that we needed a few different plugins to achieve the same as
the single plugin but each performing a different function.

For this I have removed CoC and ALE to try simplify this all.