Move to treesitter supported syntax theme.
For this I found out from taking on an older theme I was forcing tmux to
use 256 colours which was restricting the colour options. After removing
that the colours are far better.

I played around between these two for a while. I am keeping the sonokai
as I do not mind it but just going with the smallest change possible to
start off with. Despite the colour theme actually being somewhat
different than the previous tokyonight.