Add cookbook script.
This was a very basic script to just keep track of those common
commands or actions that I perform but not enough that I remember them
exactly. For now it is a simple script but has the potentnial to be
expanded on a bit to manage the creation of entries, etc.