Tmux improvements and changing task manager
tmux now can now easily resize the windows with vim-like commands. Also
made a few short cuts to jump to a note scratch pad and dot files.

I have changed my task manager over to task warrior as wanted to give it
a try and it has been pretty good. In particular it is easier to track
priorities and the print out is a bit nicer. It also has CLI
completions.

Minor changes to vim by moving to using maintained tree sitter syntaxes.