Remove zoxide for fish z plugin.
A few reasons for changing back. First being that I found in the fish
plugin I have tab completion which just searches from the stored list
whereas in zoxide I needed to explicitly start the request with "zi".

I can make the same functionality using the arguments that z already has
and simply piping the results in to fzf.