Changing up the way I manage tmux
Using sessions now for each workspace of work I am using. Previously I
would be making new windows but would get lost between which ones were
related to builds of their source, etc.

Now can group all these by a single session which can be chossen by
tmux-sessionizer (thanks Primagen).

Also have made it so that the name of the terminal is helpful for seeing
which location it is open for and which session.