Fix gd script to run from any location
https://stackoverflow.com/questions/53675286/how-does-git-interpret-a-colon-followed-by-a-path

The gd script preview was failing whenever it was run outside of the
root directory. The reason for this was that the path was being given
was relative to the root directory but when the preview command executed
it was calling for a path that did not exist. From the above link we can
prefix a pathspec with :/ and git will interpret that from the root
directory, matching the expected format.