Lots of small tweaks here.
Minor improvements in some of the scripts with ease of use and
information. Have removed the kotlin lsp script as this is now managed
through a neovim plugin. Also have removed the good morning script as
this is now only calling the upgrade script, if there are more elements
later I can easily bring it back.

Another small change is the task action is able to loop to I do not need
to create a new window each time and can actually just be used for
journalling as well.