Add fish plugin and fetch when open repo.
There is no LSP for fish shell so removing the polyglot plugin has made
editing fish files a bit difficult. Adding this plugin brings back the
nice editing experience.

I have also updated the fundamental nv script to now always fetch when
opening the editor. Reason for this was I have had time when I go to
make changes and miss that the repo is not updated. I have this checked
at the start of the day for common repositories but this will at least
catch it for all repositories now.