Several fish updates and some new nvim plugins.
Fixed the _is_work function after it actually being tested at work. Have
also removed a lot of the extra logging that was being printed in the
various scripts used daily.

A major change is in the repo-stat script which now has some nicer print
out and is using the in built git status output instead of parsing it
ourselves.

Added neoformat plugin as need a better formatter and changed the update
for tree-sitter to be synchronous as it keeps getting skipped in the
update script.