Improve the cmp experience and add a few nvim plugins
Improved the configuration based on video for cmp. Have also added some
new sources for it and some pretty formatting.
Have also added rhubarb for working with Github and gv to easily see the
changes in a file over git.
Last bit was setting up the snippets engine with cmp.

https://github.com/tjdevries/config_manager/blob/master/xdg_config/nvim/after/plugin/completion.lua