Starting of the repo script suite.
Wanting to have a nice tool to manage my common workflows within a
repository. The goal is to have a subcommand for each of the common
tasks that way I only need to remember the one command.

Also thinking of putting others under the `dev` command which affect
multiple repositories or the general development environment.