Fix font configuration and remove alias to rust versions
The aliases to the rust versions of common tools were causing some
issues as they were not exactly the same APIs.

The font configuration was rendering things slightly incorrectly in
regards to powerline fonts and icons. Removing mono and just providing
the fallback without the render style fixed that.

Also added a colouring plugin to visualise colours in nvim