Start to work back to FZF
I am really enjoying the use of fzf in all other elements of my terminal
so I thought it would be good to unify it. Also I was hitting a couple
of issues with telescope that I didn't enjoy that I just had out of the
box with FZF. The main one being history of terms.

I also added in a leap extension to have smart f/t searching so that it
will not matter if I map ;/, to anything as I won't need them anymore.