Add formatting tool and general tidy up of unused elements. Added cmp sorting
Did a tidy up of the TabNine cmp integration as it was more complicated than needed and it now uses the same colorscheme.

Added conform which is a nice formatter plugin that allows you to use both an external formatter or fall back to the LSP formatter which
is exactly what I was wanting.

Removed some of the rust specific set up as I am not really using that anymore.

Some small improvements with FZF to include hidden files and a litle script to close all pop up windows when focus is lost.

One of the biggest wins was the improvement on cmp sorting which I have had issues with in the past. This set up has done me well so
far after a few weeks of playing around with it but may still need some more tweaks.