Bunch of small nvim tweaks
Improve the cmp usage by mainly just simplifying it down to what is currently in the README

Added a few useful FZF keymaps to use the basic grep as precedence and blob when I want. Also allow for word grep in visual mode

I have started using my 'change' script again which is really handy so I removed the InsertLeave for save as it triggers it too much.

Dressing also has the option of using normal vim modes in the input window, which is great so I can use the usual jumps around when renaming.